Create BNBTiger
Coin tiger BNBTiger Shib killer
We've had enough of scams in the market! These projects reduce people's trust, lead to suspicion and panic, and finally let one project die, and there will be no great projects like Shib!
The goal of the birth of BNBTiger is to rebuild the security and trust of the market. Like Shib, we completed another 26000 times task. Therefore, we launched a huge amount of 10,000,000,000,000,000,000,000,000 tokens, destroyed 50%, added 50% liquidity and launched fairly! Let's start with 0. This is a new social experiment. We want to see how many 0 can be killed by real trust! Our ultimate goal is to eliminate all 0 and surpass Shib's historical record!